+++ doc/source/changelog.rst
@@ -12,6 +12,10 @@ v3.0.1
 
 * Fixed bug with labels of ``plot_implicit``.
 
+* Fixed bug with difficult to render labels on Matplotlib. If Matplotlib
+  detects an error while parsing legend's entries, the plot won't show
+  the legend.
+
 
 v3.0.0
 ======